chore: Page Pane visual revisions
