+++ app/client/src/pages/Editor/IDE/EditorPane/AddWidgets.tsx
@@ -1,12 +1,13 @@
 import React, { useCallback } from "react";
-import { Flex, Text, Button } from "design-system";
+import { Flex } from "design-system";
 import { useSelector } from "react-redux";
 
 import history from "utils/history";
 import { getCurrentPageId } from "@appsmith/selectors/entitiesSelector";
 import WidgetSidebarWithTags from "pages/Editor/WidgetSidebarWithTags";
 import { widgetListURL } from "@appsmith/RouteBuilder";
-import { createMessage, PAGES_PANE_TEXTS } from "@appsmith/constants/messages";
+import { PAGES_PANE_TEXTS } from "@appsmith/constants/messages";
+import SegmentAddHeader from "./components/SegmentAddHeader";
 
 const AddWidgets = () => {
   const pageId = useSelector(getCurrentPageId) as string;
@@ -17,28 +18,10 @@ const AddWidgets = () => {
 
   return (
     <>
-      <Flex
-        alignItems="center"
-        borderBottom={"1px solid var(--ads-v2-color-border)"}
-        justifyContent="space-between"
-        px="spaces-4"
-        py="spaces-2"
-      >
-        <Text
-          className="overflow-hidden overflow-ellipsis whitespace-nowrap"
-          color="var(--ads-v2-color-fg)"
-          kind="heading-xs"
-        >
-          {createMessage(PAGES_PANE_TEXTS.widgets_create_tab_title)}
-        </Text>
-        <Button
-          isIconButton
-          kind={"tertiary"}
-          onClick={closeButtonClickHandler}
-          size={"sm"}
-          startIcon={"close-line"}
-        />
-      </Flex>
+      <SegmentAddHeader
+        onCloseClick={closeButtonClickHandler}
+        titleMessage={PAGES_PANE_TEXTS.widgets_create_tab_title}
+      />
       <Flex flexDirection="column" gap="spaces-3" height="calc(100vh - 310px)">
         <WidgetSidebarWithTags isActive />
       </Flex>
