+++ app/client/src/pages/Editor/IDE/EditorPane/EditorPaneSegments.tsx
@@ -1,6 +1,5 @@
 import React from "react";
-import { Button, Flex, SegmentedControl } from "design-system";
-import { createMessage, PAGES_PANE_TEXTS } from "@appsmith/constants/messages";
+import { Flex } from "design-system";
 import { Switch, useRouteMatch } from "react-router";
 import { SentryRoute } from "@appsmith/AppRouter";
 import { QueriesSection } from "./QueriesSection";
@@ -18,91 +17,15 @@ import {
 import { SAAS_EDITOR_API_ID_PATH } from "../../SaaSEditor/constants";
 import { JSSection } from "./JS_Section";
 import { WidgetsSection } from "./WidgetsSection";
-import { useSelector } from "react-redux";
-import { getCurrentPageId } from "@appsmith/selectors/entitiesSelector";
-import history, { NavigationMethod } from "utils/history";
-import {
-  globalAddURL,
-  jsCollectionListURL,
-  queryListURL,
-  widgetListURL,
-} from "@appsmith/RouteBuilder";
-import { EditorEntityTab } from "entities/IDE/constants";
 import EntityProperties from "pages/Editor/Explorer/Entity/EntityProperties";
-import { useCurrentEditorState } from "../hooks";
+import SegmentedHeader from "./components/SegmentedHeader";
 
 const EditorPaneSegments = () => {
-  const pageId = useSelector(getCurrentPageId);
   const { path } = useRouteMatch();
-  const { segment } = useCurrentEditorState();
-  /**
-   * Callback to handle the segment change
-   *
-   * @param value
-   * @returns
-   *
-   */
-  const onSegmentChange = (value: string) => {
-    switch (value) {
-      case EditorEntityTab.QUERIES:
-        history.push(queryListURL({ pageId }), {
-          invokedBy: NavigationMethod.SegmentControl,
-        });
-        break;
-      case EditorEntityTab.JS:
-        history.push(jsCollectionListURL({ pageId }), {
-          invokedBy: NavigationMethod.SegmentControl,
-        });
-        break;
-      case EditorEntityTab.UI:
-        history.push(widgetListURL({ pageId }), {
-          invokedBy: NavigationMethod.SegmentControl,
-        });
-        break;
-    }
-  };
-
-  const onAddButtonClick = () => {
-    history.push(globalAddURL({ pageId }));
-  };
 
   return (
     <Flex flexDirection="column" gap="spacing-2" overflow="hidden">
-      <Flex
-        alignItems="center"
-        className="ide-pages-pane__header"
-        gap="spaces-2"
-        justifyContent="space-between"
-        padding="spaces-2"
-      >
-        <SegmentedControl
-          isFullWidth
-          onChange={onSegmentChange}
-          options={[
-            {
-              label: createMessage(PAGES_PANE_TEXTS.queries_tab),
-              value: EditorEntityTab.QUERIES,
-            },
-            {
-              label: createMessage(PAGES_PANE_TEXTS.js_tab),
-              value: EditorEntityTab.JS,
-            },
-            {
-              label: createMessage(PAGES_PANE_TEXTS.ui_tab),
-              value: EditorEntityTab.UI,
-            },
-          ]}
-          value={segment}
-        />
-        <Button
-          className={"t--add-editor-button"}
-          isIconButton
-          kind="primary"
-          onClick={onAddButtonClick}
-          size="sm"
-          startIcon="add-line"
-        />
-      </Flex>
+      <SegmentedHeader />
       <EntityProperties />
       <Flex
         className="ide-pages-pane__content"
